Rename `config_` to `prototype_` which makes more sense
