+++ src/prototype_creep_harvest.js
