+++ src/prototype_creep_resources.js
