+++ src/prototype_room.js
