+++ src/prototype_roomPosition.js
