+++ src/prototype_room_external.js
