+++ src/prototype_room_fight.js
