+++ src/prototype_room_init.js
