+++ src/prototype_room_memory.js
