+++ src/prototype_room_mineral.js
