+++ src/prototype_room_utils.js
