+++ src/prototype_room_wallsetter.js
