+++ src/prototype_string.js
