+++ src/require.js
@@ -3,32 +3,32 @@ require('config_logging');
 require('config_brain_memory');
 require('config_brain_nextroom');
 require('config_brain_squadmanager');
-require('config_creep');
-require('config_creep_clean');
-require('config_creep_fight');
-require('config_creep_resources');
-require('config_creep_harvest');
-require('config_creep_mineral');
-require('config_creep_nextroom');
-require('config_creep_routing');
-require('config_roomPosition_structures');
-require('config_room');
-require('config_room_attack');
-require('config_room_basebuilder');
-require('config_room_cached');
-require('config_room_controller');
-require('config_room_creepbuilder');
-require('config_room_defense');
-require('config_room_fight');
-require('config_room_market');
-require('config_room_memory');
-require('config_room_mineral');
-require('config_room_my');
-require('config_room_external');
-require('config_room_routing');
-require('config_room_utils');
-require('config_room_wallsetter');
-require('config_string');
+require('prototype_creep');
+require('prototype_creep_clean');
+require('prototype_creep_fight');
+require('prototype_creep_resources');
+require('prototype_creep_harvest');
+require('prototype_creep_mineral');
+require('prototype_creep_nextroom');
+require('prototype_creep_routing');
+require('prototype_roomPosition_structures');
+require('prototype_room');
+require('prototype_room_attack');
+require('prototype_room_basebuilder');
+require('prototype_room_cached');
+require('prototype_room_controller');
+require('prototype_room_creepbuilder');
+require('prototype_room_defense');
+require('prototype_room_fight');
+require('prototype_room_market');
+require('prototype_room_memory');
+require('prototype_room_mineral');
+require('prototype_room_my');
+require('prototype_room_external');
+require('prototype_room_routing');
+require('prototype_room_utils');
+require('prototype_room_wallsetter');
+require('prototype_string');
 require('role_atkeeper');
 require('role_atkeepermelee');
 require('role_autoattackmelee');
